Actualización de modelos

+++ XTecDigital/Models/Archivo.cs
@@ -8,7 +8,7 @@ namespace XTecDigital.Models
         public int Id { get; set; }
         public int IdCarpeta { get; set; }
         public string Nombre { get; set; }
-        public decimal Tamanio { get; set; }
+        public int Tamanio { get; set; }
         public DateTime? FechaCreacion { get; set; }
 
         public virtual Carpeta IdCarpetaNavigation { get; set; }
